Financial Report Bank Companies
Sales figures from financial reports for five public companies in an industry. They were asked to analyze what factors caused sales fluctuations in the five companies based on the financial reports, and then design a dashboard visualization from the processed data to analyze how sales trends in the industry based on the interpretation of the dashboard visualization results.